requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.54.0
httpx==0.27.0
psycopg2-binary==2.9.9
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-search-results==2.4.2
markdown==3.5.1
